bad contests + USACO stuff 2/3/19

+++ USACO/Contest/January2019/Post-Contest Solutions/shortcut.cpp
@@ -18,7 +18,7 @@ struct CompareState{
 
 struct Edge{
    int to;                                //other vertex
-   int w;                                 //weight
+   long long w;                                 //weight
 };
